Power up rotations
Updated the colliders on the power-ups to be box colliders instead of
capsule.  Probably not a huge difference but my OCD really likes it a
lot better cause it fits the object MUCH better.  Also added a rotate
option to the script so that the power-ups are rotating.  Need to
figure why the one for the cross is off center though.